pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dagster-unified-platform"
version = "2.1.0"
description = "Unified MLOps/CVOps/LLMOps platform with multi-source support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Data Team", email = "data@example.com"}
]
keywords = ["dagster", "mlops", "cvops", "llmops", "lakehouse", "iceberg", "feast"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core Dagster
    "dagster>=1.9.0,<1.10.0",
    "dagster-webserver>=1.9.0,<1.10.0",
    "dagster-postgres>=0.25.0,<0.26.0",
    
    # Data Infrastructure
    "lakefs-sdk>=1.71.0",
    "trino>=0.328.0",
    "boto3>=1.35.36",
    "pyiceberg[pyarrow,duckdb]>=0.8.1",
    "pynessie>=0.67.0",
    
    # Database Connectors
    "psycopg2-binary>=2.9.0",  # PostgreSQL
    "PyMySQL>=1.1.1",          # MySQL
    "pyodbc>=5.0.0",           # SQL Server
    "sqlalchemy>=2.0.0",
    
    # Data Processing
    "pandas>=2.2.0",  # pandas 2.2+ supports numpy 2.x
    "pyarrow==17.0.0",
    "duckdb==0.10.0",
    
    # API & Requests
    "requests==2.32.3",
    "httpx>=0.25.0",
    
    # Configuration & Environment
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Feature Store & ML
    "mlflow==3.6.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    
    # Utilities
    "tenacity>=8.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Feast Feature Store Dependencies
feast = [
    "feast[trino]==0.57.0",
    "psycopg[binary]>=3.1.0",  # PostgreSQL (SQLAlchemy 2.x / Feast)
    "redis==5.0.1",
    "fastapi==0.108.0",
    "uvicorn[standard]==0.34.0",  # Feast 0.57.0 requires >=0.30.6,<=0.34.0
]

# Computer Vision Dependencies (Enhanced for CVOps)
cv = [
    # YOLO & Deep Learning
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # Image Processing
    "opencv-python-headless>=4.8.0",  # Headless for servers
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    
    # Image utilities
    "imagehash>=4.3.0",        # Perceptual hashing for deduplication
    "ffmpeg-python>=0.2.0",    # Video processing
    
    # API for inference service
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6", # File uploads
    
    # Kafka for streaming detections
    "confluent-kafka>=2.2.0",
]

# LLM Dependencies
llm = [
    "sentence-transformers>=2.2.0",
    "transformers>=4.30.0",
    "langchain>=0.1.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
]

# Development Dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

# Notebook Dependencies
notebook = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "papermill>=2.5.0",  # Notebook automation
]

# Production Monitoring
monitoring = [
    "prometheus-client>=0.17.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "evidently>=0.4.0",  # Drift detection
]

# Streaming
streaming = [
    "kafka-python>=2.0.0",
    "confluent-kafka>=2.3.0",
]

# Webhook server (Flask-based)
webhook = [
    "flask>=2.3.0",
    "gunicorn>=21.0.0",
]

# All dependencies
all = [
    "dagster-unified-platform[feast,cv,llm,dev,notebook,monitoring,streaming,webhook]",
]

[project.urls]
Homepage = "https://github.com/yourorg/dagster-unified-platform"
Documentation = "https://github.com/yourorg/dagster-unified-platform#readme"
Repository = "https://github.com/yourorg/dagster-unified-platform"
Issues = "https://github.com/yourorg/dagster-unified-platform/issues"

[tool.dagster]
module_name = "src.pipelines"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"src.pipelines" = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src", "src.api", "src.pipelines", "src.streaming"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (require services)",
    "e2e: End-to-end tests (full pipeline)",
    "slow: Slow-running tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["src/api", "src/pipelines", "src/streaming"]
omit = [
    "**/tests/*",
    "**/__pycache__/*",
    "**/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 0
show_missing = true